+++ melspec-to-video.py
@@ -1,10 +1,8 @@
 import argparse
-import glob
 import logging
 import math
 import os
 import subprocess
-import sys
 import threading
 import time
 from typing import Any, Final, List
@@ -15,11 +13,9 @@ import matplotlib.pyplot as plt
 import numpy as np
 import psutil
 import soundfile as sf
-import yaml
 from PIL import Image, ImageDraw, ImageFont
 from tqdm import tqdm
-from datetime import datetime
-import json
+import utils
 
 # Mel Scale Spectrogram Video from Audio
 
@@ -128,13 +124,6 @@ def monitor_memory_usage(interval=1):
         if max_mem < memory_usage:
             max_mem = memory_usage
         time.sleep(interval)
-
-
-def load_config(config_path):
-    """Load YAML configuration file."""
-    with open(config_path, "r") as file:
-        config = yaml.safe_load(file)
-    return config
 
 
 def calculate_buffer(
@@ -637,33 +626,6 @@ def profile_audio(config: dict[str, Any], args: Any) -> dict[str, Any]:
     }
 
 
-def update_config(config, config_path, audio_path, profile):
-    # Ensure the "audio_files" section exists
-    if "audio_files" not in config:
-        config["audio_files"] = {}
-
-    # Update the specific audio file's profile within "audio_files"
-    if audio_path not in config["audio_files"]:
-        config["audio_files"][audio_path] = {}
-
-    # Update the profile for the specific audio file
-    config["audio_files"][audio_path]["max_power"] = float(profile["max_power"])
-    config["audio_files"][audio_path]["sample_rate"] = profile["sample_rate"]
-    config["audio_files"][audio_path]["sample_count"] = profile["sample_count"]
-
-    # Try to save the updated configuration back to the YAML file
-    try:
-        with open(config_path, "w") as file:
-            yaml.safe_dump(config, file)
-        logging.info("Configuration updated successfully.")
-    except Exception as e:
-        logging.error(f"Failed to update configuration: {e}")
-        # Consider whether to terminate the program or handle the error differently
-        raise  # Re-throw the exception for the caller to handle
-
-    return config
-
-
 def generate_spectrograms(
     config: dict[str, Any],
     args: argparse.Namespace,
@@ -778,83 +740,11 @@ def generate_spectrograms(
         progress_bar.update(duration_secs)
         count += 1
     progress_bar.close()
-    project_data = update_project_data(
+    project_data = utils.update_project_data(
         project_folder, "images_metadata", images_metadata
     )
 
     return True
-
-
-def generate_folder_name(base_name):
-    timestamp = datetime.now().strftime("%y%m%d%H%M%S")
-    folder_name = f"{base_name}-{timestamp}"
-    return folder_name
-
-
-def create_project_folder(base_path, folder_name):
-    project_path = os.path.join(base_path, folder_name)
-    if not os.path.exists(project_path):
-        os.makedirs(project_path)
-    return project_path
-
-
-def initialize_project_data(project_folder, fullpath):
-    directory_path = os.path.dirname(fullpath)
-    filename = os.path.basename(fullpath)
-    json_filename = os.path.join(project_folder, "project_data.json")
-
-    initial_data = {
-        "audio_metadata": {
-            "source_audio_path": directory_path,
-            "source_audio_filename": filename,
-            "max_power": None,
-            "sample_rate": None,
-            "sample_count": None,
-        },
-        "images_metadata": [],
-    }
-    # Write the project data to the JSON file
-    with open(json_filename, "w") as json_file:
-        json.dump(initial_data, json_file, indent=4)
-
-    return initial_data
-
-
-def load_project_data(project_folder):
-    json_filename = os.path.join(project_folder, "project_data.json")
-
-    if os.path.exists(json_filename):
-        with open(json_filename, "r") as json_file:
-            return json.load(json_file)
-    else:
-        print("Project data file not found.")
-        return None
-
-
-def update_project_data(project_folder, key, new_data):
-    json_filename = os.path.join(project_folder, "project_data.json")
-    print(f"json_filename: {json_filename}")
-    # Ensure the project data file exists
-    if not os.path.exists(json_filename):
-        print("Project data file not found. Initial data setup may be required.")
-        return
-
-    # Load the existing project data
-    with open(json_filename, "r") as json_file:
-        project_data = json.load(json_file)
-        print(f"current project data : {project_data}")
-    # Check if the key exists and if it points to a list, append the new data
-    if key in project_data and isinstance(project_data[key], list):
-        project_data[key].append(new_data)
-    else:
-        # For non-list data or new keys, update or set the value directly
-        project_data[key] = new_data
-
-    print(f"updated project data : {project_data}")
-    # Write the updated project data back to the file
-    with open(json_filename, "w") as json_file:
-        json.dump(project_data, json_file, indent=4)
-    return project_data
 
 
 def main():
@@ -914,7 +804,7 @@ def main():
 
     args = parser.parse_args()
 
-    config = load_config(args.config)
+    config = utils.load_config(args.config)
 
     full_path = os.path.abspath(args.input)
     directory_path = os.path.dirname(full_path)
@@ -923,11 +813,11 @@ def main():
 
     print(f"The full path of the input file is: {full_path}")
 
-    project_folder = generate_folder_name(basename)
-    create_project_folder(directory_path, project_folder)
+    project_folder = utils.generate_project_folder_name(basename)
+    utils.create_folder(directory_path, project_folder)
     print(f"Project folder : {project_folder}")
 
-    project_data = initialize_project_data(project_folder, full_path)
+    project_data = utils.initialize_project_data(project_folder, full_path)
     print(f"project data init: {project_data}")
     ### PASS 1
     ## if these are missing recalculate and save to config file
@@ -940,7 +830,9 @@ def main():
             config, args
         )  # Assuming profile_audio is adjusted to not require 'config'
         print(f"Profile : {profile}")
-        project_data = update_project_data(project_folder, "audio_metadata", profile)
+        project_data = utils.update_project_data(
+            project_folder, "audio_metadata", profile
+        )
 
     print(f"project data : {project_data}")
     ### PASS 2
